+++ grand_trade_auto/database/database_meta.py
@@ -10,6 +10,11 @@ Module Attributes:
 (C) Copyright 2020 Jonathan Casey.  All Rights Reserved Worldwide.
 """
 from abc import ABC, abstractmethod
+import logging
+
+
+
+logger = logging.getLogger(__name__)
 
 
 
@@ -19,21 +24,20 @@ class Database(ABC):
     will subclass this, but externally will likely only call the generic methods
     defined here to unify the interface.
 
+    This serves as a base class for other database systems, so will consume
+    any final kwargs.
+
     Class Attributes:
       N/A
 
     Instance Attributes:
-      env (str): The run environment type valid for using this database.
-      cp_db_id (str): The id used as the section name in the database conf.
+      _env (str): The run environment type valid for using this database.
+      _cp_db_id (str): The id used as the section name in the database conf.
         Will be used for loading credentials on-demand.
-      cp_secrets_id (str): The id used as the section name in the secrets
+      _cp_secrets_id (str): The id used as the section name in the secrets
         conf.  Will be used for loading credentials on-demand.
-      host (str): The host URL.
-      post (int): The port number on that host for accessing the database.
-      database (str): The database to open.
     """
-    def __init__(self, env, cp_db_id, cp_secrets_id,
-            host, port, database):             # pylint: disable=unused-argument
+    def __init__(self, env, cp_db_id, cp_secrets_id, **kwargs):
         """
         Creates the database handle.
 
@@ -43,13 +47,14 @@ class Database(ABC):
             Will be used for loading credentials on-demand.
           cp_secrets_id (str): The id used as the section name in the secrets
             conf.  Will be used for loading credentials on-demand.
-          host (str): The host URL.
-          post (int): The port number on that host for accessing the database.
-          database (str): The database to open.
         """
         self._env = env
         self._cp_db_id = cp_db_id
         self._cp_secrets_id = cp_secrets_id
+
+        if kwargs:
+            logger.warning('Discarded excess kwargs provided to'
+                    + f' {self.__class__.__name__}: {", ".join(kwargs.keys())}')
 
 
 
@@ -101,7 +106,7 @@ class Database(ABC):
             appears as the section header in the secrets_cp.
 
         Returns:
-          db_handle (Database<>): The Database<> object created and loaded from
+          db (Database<>): The Database<> object created and loaded from
             config, where Database<> is a subclass of Database (e.g. Postgres).
         """